requirements.txt
# Core dependencies for plot_quokka
yt>=4.0.0
unyt>=2.9.0
numpy>=1.20.0
matplotlib>=3.5.0
PyYAML>=5.0

# Optional: for web frontend
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
